Fix new E2E pipeline (#2966)
A previous change in EdgeAgent (https://github.com/Azure/iotedge/commit/11d4e3a66f1986aa47c4ee8f4fdb70550a435d3a) allowed EdgeAgent to be restarted when the config is updated. This, however, broke the new e2e test pipeline.

The change in EA stores environment variables as labels on the new EA container before running it and deleting the old container. However, this change requires that both versions of EA know about the change to work.

So EA version with the change (let's call it change x) is incompatible with an EA version without change x.

In the new e2e pipeline, we start up bootstrapping 1.0 -> which pulls 1.0.9.1 currently. 1.0.9.1 does not have change x, and will not be receiving change x.  But in development and in the pipeline, the bootstrap pulls down the latest version of EA in master, so this test will fail since these are incompatible. 

This PR defines a new bootstrap image for EA for the pipeline. This is temporary. Once 1.0 knows about the change OR once we make the EA image backwards compatible, we can remove this fix.

This PR also contains a change to the E2E module comparison logic when a module gets started up. If the CreateOptions is filled out in an image with the prefix 'net.azure-devices.edge' then all of the CreateOptions will be ignored in comparison of reported and desired properties. Otherwise the test will never verify that EA module got deployed, since EA (after change x) "magically" inserts environment variables into the Labels portion of CreateOptions. The test doesn't specify these to be filled out in DesiredProperties, so the test doesn't know about them. 

Ignoring the entirety of CreateOptions might be too crude, so if anyone has an idea to make it more specific we can change that (maybe we could make it for just EA and hardcode EA name in the comparison to check first). I think it is fine for now, though, because that prefix shouldn't get into labels any other way, so no other module should run into this issue. And EA never changes CreateOptions in any of the new E2E tests, so it should be safe for now.